Update outdated leaderboard on request
