--- app/crons/script.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-cd $1
-npx ts-node app/crons/leaderboard.ts
